setup recyclerview

+++ app/src/main/java/com/example/shoplist/domain/GetShopItemUseCase.kt
@@ -2,6 +2,6 @@ package com.example.shoplist.domain
 
 class GetShopItemUseCase(private val shopListRepository: ShopListRepository) {
     fun getItemId(shopItemId: String): ShopItem {
-        return shopListRepository.getItemId(shopItemId)
+        return shopListRepository.getItemId(shopItemId.toInt())
     }
 }
